+++ tomcat/src/main/java/org/apache/coyote/handler/mapping/login/LoginMapping.java
@@ -1,7 +1,8 @@
-package org.apache.coyote.handler.mapping;
+package org.apache.coyote.handler.mapping.login;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.coyote.handler.mapping.HandlerMapping;
 import org.apache.coyote.http.common.HttpBody;
 import org.apache.coyote.http.common.HttpHeader;
 import org.apache.coyote.http.common.HttpHeaders;
