+++ tomcat/src/main/java/org/apache/coyote/handler/mapping/register/RegisterMapping.java
@@ -1,7 +1,8 @@
-package org.apache.coyote.handler.mapping;
+package org.apache.coyote.handler.mapping.register;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.coyote.handler.mapping.HandlerMapping;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.response.HttpResponse;
